Fix #45: max (renewable) ticket life is negative
Since the negative flag in the helper function pykadmin_timestamp_as_deltastr
was never initialized to zero, using it in a comparison reads random values off
the stack, which are more likely to be non-zero than zero. Hence, checking
whether the delta was negative would return true even in cases where the delta
was not negative.